"10:30am. I last heard from him 9 days ago. It is even longer than I thought. Those dependent langs really took a lot more of my time than I expected.
11:15am.

///

Hello again. Let me start by apologizing for leaving for weeks at a time. For past month I've been trying to get Lean - and since our last contact Agda into working shape, but I've given up on that. I am a fairly persistent person so it took me this long to realize that trying to prove the paper theorems formally is not the best approach to take if I want to properly cover the papers that I want. 

Maybe you've heard the rumors, and if so, I can confirm them. Formal theorem proving is really, really hard. I had this epiphany when I tried implementing the [insert](https://github.com/mrakgr/CFR-In-Agda/blob/54bbc48165a803aa5534509fa0094e9b499cbb93/AssociationList.agda) function in Agda. If it were merely the implementation, this would've been a 5m job in F#, but having to maintain the proof terms that the lists has no repeating elements turned it into a 4h job in Agda. The work had to be done over two days and I was quite exhausted by the end of it.

More to the point, in order to do a formal proof I actually need to know the necessary proof steps down to the tiniest detail if I want to make it.

I've been hoping that the insight would come to me once I pick up the tools and do the implementation, but even as far as yesterday I still had absolutely no idea how to prove eq 9. So what am I to do?

In the years that I've been aware of it, I wrote [randomized testing](https://fscheck.github.io/FsCheck/QuickStart.html) off as a gimmick, but after months of experience in formal proving, I am starting to reconsider its true worth. To be sure, writing the generators for the tests is some work, but it is still far easier than doing a formal proof. Thinking back to my experiences of going through Software Foundations and the rest, any kind of formal proof one can do constructively - one can also do an approximate verification much more easily using randomized testing.

So this is what I am going to be using to go through troublesome papers from here on out.

---

Now to the issue at hand, yesterday I've done an implementation of the regret calculation and implemented eq 9 from the paper directly.

https://github.com/mrakgr/CFR-In-Fsharp/tree/bfc4318c3b88bb602d2bcb0ce7d597fc23a1c324/CFR%20In%20Fsharp/Eq%209  

I've done two different version in `eq9.fsx` and `eq9'.fsx` and both of them fail, so based on that I can only conclude that eq 9 is invalid. This applies to equality (=?), inequality (<=) and inequality (>=).

If you feel like disputing this, it might be good to do a review and on my end, I am willing to explain everything that is needed until we come to an agreement. To be far, I can't 100% rule out the possibility that there is an error in my own code, though I am not suspicious of that at the moment.

If you are familiar with any FP language, then I do not think you will have too much trouble understanding the F# code, and even if you aren't - unlike Agda and Lean, F# is actually a fairly simple language. On Windows, it is trivial to install as it comes with the free Visual Studio. On Linux I am not sure, still MS is making strides to make .NET work there.

///

Here it is. Since I posting his emails, it is only fair that I post mine.

Now that I've done this, what comes next?

Well, I should test the full theorem directly. I think that even if the proofs are invalid the idea that full regret <= sums of positive immediate regrets should be valid.

11:25am. Beyond that I should track down how regret bounds are calculated.

11:30am. After that comes the other papers. In VR-MCCFR paper there was also that suspicious step which I want to go through.

11:35am. At any rate this is good. After essentially doing nothing direct to my interest for not just the past month, but maybe the entirety of 2019, I've finally found something that I can grasp.

My skillset right now is actually extremely well suited for doing these kinds of investigations. Compared to doing formal proofs, doing these generators is a piece of cake. I have years of functional programming experience to fall back on here.

And maybe, things going down this course is the ideal outcome of all as it will align my desires with the requirements of reality. Maybe if I do my studies in this fashion I will discover something related to AI merely from trying to automate testing.

This whole thing started when I made that post on the PL sub. At point I got various answers, but the correct one was missing - randomized testing.

11:40am. Let me take a break here, I'll resume after that. Maybe with this, I'll be able to go through the following years in a more leisurely fashion. The last few years were pretty brutal in terms of workload and I have nothing concrete to show for it. All these unknowns were making me work way too hard.

But from here on out, I am going to get a handle on all the fields of math that I need."