"2:35pm.
```
The quantities in (8) inside the right-most pair of parentheses is c - b. This is because succ defines a distribution over I' (which btw is a function of the ratio of reach probabilities between I and I', which I agree is not super clear in the proof -- I only know this after years of thinking about it!) and because the transitions between I and I' are weighted properly (i.e. according to those probabilities).

Now let's concentrate on a - d. This term is actually zero, so you're adding nothing, hence the equality. The reason is very similar to c - b: the only difference is the policy is \sigma^t | I -> a instead of \sigma^t. But that doesn't matter because the logic still holds.

Since that term is zero, it doesn't matter that you have max_{a \in A(I)} in front of T because it's true for any subset of actions on any time step, so it's trivially true for a specific single action over all time steps.
```

3:15pm. I had to take a break for a while. I've been deep in thought during this time.

I got it. I now understand the proof in full. And I understand the missing step.

```
        action_max o id (fun i o_a ->
            u o_a tree - u o tree +
            // According to definition of succ which only considers opponent reach probabilities,
            // I skip multipling by current player reach probabilities.
            // Since perfect recall assures no cycles in infosets `u o' branch = u (action_at i o' id) branch` holds.
            u o' branches.[i] - u o branches.[i]
            )
```

Note that here, `u o_a tree` is in fact equal to `u o branches.[i]`.

Lanctot was right! Only I could not see it because all of the noise in my head. But it is really obvious when written out like this.

3:25pm. Damn, I do feel like taking a break. While I was thinking I was considering doing a formal proof in Agda, but forget that. I am going to quickcheck it all the way.

Actually, before I do anything else, how would the recursion work.

Ah, I should be able to prove that `max 0.0 (a + max 0.0 b) <= max 0.0 a + max 0.0 b` and use that fact to flatten the tree.

3:35pm. Ah, let me do it. Let me prove all the steps in one file from the start to finish.

Holy shit is VS annoying with its renaming bugs.

3:50pm. I decided to report this issue.

Let me get on with what I wanted to do.

...These folder issues are driving me crazy. I can't even add items properly. Maybe I should figure out how to set up VS Code and ditch the new VS.

3:55pm. Right now everything is broken in the solution.

4pm. Ah, I figured it out. It adds the stupid items out of order.

The IDE has a tendency to add the files in the wrong order when adding to folders.

```
    <None Include="Theorem 3\count_and_pick.fsx" />
    <None Include="Theorem 3\Core.fsx" />
    <None Include="Theorem 3\Simplified\Core.fsx" />
    <None Include="Theorem 3\Simplified\example.fsx" />
    <None Include="Theorem 3\Simplified\Lemma 5\eq 13.fsx" />
    <None Include="Theorem 3\Simplified\Lemma 5\eq 9.fsx" />
    <None Include="Theorem 3\Simplified\Lemma 5\example.fsx" />
    <None Include="Theorem 3\Simplified\Lemma 5\Core.fsx" />
    <None Include="Theorem 3\Simplified\Lemma 5\readme.md" />
    <None Include="Theorem 3\Simplified\readme.md" />
```

Just now I had the files in this order in the project file. I had to fix it by changing it manually to the following otherwise it would complain that a directory would get rendered multiple times.

```
    <None Include="Theorem 3\count_and_pick.fsx" />
    <None Include="Theorem 3\Core.fsx" />
    <None Include="Theorem 3\Simplified\Core.fsx" />
    <None Include="Theorem 3\Simplified\example.fsx" />
    <None Include="Theorem 3\Simplified\readme.md" />
    <None Include="Theorem 3\Simplified\Lemma 5\eq 13.fsx" />
    <None Include="Theorem 3\Simplified\Lemma 5\eq 9.fsx" />
    <None Include="Theorem 3\Simplified\Lemma 5\example.fsx" />
    <None Include="Theorem 3\Simplified\Lemma 5\Core.fsx" />
    <None Include="Theorem 3\Simplified\Lemma 5\readme.md" />
```

4:10pm. Ok, I figured out how to get around the renaming and file moving issues. It is all due to the IDE's bungling of the project file.

Forget that. It is time to bring this to a close.

Before I do that, let me stop here so I can do the chores.

After that I'll get to work on the generator.

From this point on, I will do both players and their temporal policies. I'll have to modify the utility calculations to make sure everything goes through as well.

I will do everything perfectly this time."