"8:25am. I actually woke up much earlier than this. I actually the reason for this time is uncomfortable posture that lead to back pain, though back issues is why I had an uncomfortable posture in the first place. Also I seem to have have a cold. Well, whatever. I was probably just to rigid yesterday. Both of those things will pass.
The important thing is that I do that proof. The sooner I acquire this ability, the sooner I will be complete as a programmer and will be able to make some real progress on AI.

I've been thinking about the difference between high and low functional styles, and when all is said and done, the low style is better. High functional style is about proving things to the computer, but the low style is about proving things to myself. The later suits me much better. That F# code that I wrote is high art. Agda on the other hand is just making me work a lot.

It is fine if some things do not get covered by the type system. There is plenty of structure in the code itself.

I finally did it. I finally found my math focus.

8:30am. But I am not 100% sure just yet. I only have like 4 days of experience in approximate proving. I need to make sure that what I have is solid, and for that I need to be able to cover more difficult proofs. The stuff like in the Blackwell paper and some of the other papers I've read.

In order to have confidence in this, I need to challenge myself and conquer more difficult papers.

Well, but for now, let me chill for a while. Let me see if Tog is out."