"10:25am. Let me finally start for the day.
For the first time in a while I have a lot of leeway to do what I want. Like, my general goal is to study game theory, but I have a wide berth from which to choose what I want to do next.

I want to take a break from it. I've been constantly programming for the past week, so how about I do what I said I would do in the previous entries and study Prolog for a bit. I still remember that Datalog paper which got me interested in it and I never found a way to sate that appetite.

10:35am. Rather than trying to bulldoze through everything, at this point it is best that I cast around and only start reeling once I have something.

https://www.youtube.com/channel/UCFFeNyzCEQDS4KCecugmotg
The Power of Prolog

https://edu.swi-prolog.org/

10:40am. https://www.swi-prolog.org/

I can't find much in terms of courses by Googling. Maybe I'd be best off by starting with the language itself.

https://www.reddit.com/r/prolog/

10:55pm. There are a bunch of resources on this sub.

https://www.metalevel.at/prolog

Ah, this page has it in the order it is meant to be viewed. I wasn't sure where to start just on that channel.

http://pages.cs.wisc.edu/~remzi/Naur.pdf
Programming As Theory Building

11:15am. https://www.reddit.com/r/prolog/comments/952d5v/is_prolog_still_used_today_and_is_it_still_worth/
"This post on hackernews is wonderful and inspiring:

https://news.ycombinator.com/item?id=17702946

At least if you're me, and part of the reason you got into Prolog was because you wanted to use it as an engine for creating fancy skinnable type systems."

Let me have a look. This guy is on the same page as I am it seems.

11:40am. https://www.metalevel.at/prolog/logic

Ok, I really need to make up my mind and do some proper work rather than just skirting around. I'll watch a bunch of these lectures today and start reading the '44 book by Neumann.

For now let me stop here for a bit so I can have breakfast. In fact, since I slacked earlier, I think I will just watch a video while I eat."