"1:25pm. Let me resume. It is time for more Prolog. I'll start with Data Structures.
2:35pm. Like Haskell, Prolog's strings are lists of characters. What a pity.

3:05pm. I keep being lazy and taking breaks. I still haven't gotten through data structures.

I keep thinking of how I could do type inference and I think I will give up on the typecase idea. Haskell has it right with typeclasses. Only what I will do is allow them to be local to functions and statements. That will increase their flexibility.

I'll also include extra pattern matching constructs. While I won't put in full dependent types, there is no reason not to make the compile aware of compile time values. Those will go into the type system.

3:35pm.

```
The following fragment alone already makes the most general query nonterminating in this case:
~list_list_together([], Bs, Bs) :- false.~
list_list_together([A|As], Bs, [A|Cs]) :-
        list_list_together(As, Bs, Cs).
```

This is super interesting. This issue will definitely come up during type system design.

6:10pm. https://www.metalevel.at/prolog/debugging

```
Let us briefly pause and reflect on the approach outlined above. Clearly, the method is based on the following important property:
Removing a goal can make a predicate at most more general, never more specific.
This property clearly holds for pure and monotonic Prolog code. That is a Turing complete subset of Prolog and therefore suffices to express all computations that are currently known. However, solely for the sake of argument, suppose that your application ideas are so colossal that they exceed the expressiveness of this fragment. Even if that were the case, it would still be remarkable that such a strong, general statement can be made about how changes to a program affect its meaning. Since that fragment is, in reality, as expressive as the whole language in computational terms, it is all the more remarkable!

In fact, Prolog is rather unique due to this property alone already. Compare this property to other programming languages: How do comparable changes in a Java, C++ or Haskell program affect its meaning? For example, what changes if you remove a for loop in a Java program, an if-then-else block in a C++ program, or one of the lines in a do block in a Haskell program? Does such a program even compile in all cases?

Can you make any such statement for any fragment of any programming language other than Prolog?

There are many other interesting properties that also hold for pure and monotonic Prolog code, and which can be exploited for locating mistakes in our programs. It is our hope that the availability and usefulness of such declarative methods encourage you to value and retain logical purity of your Prolog programs.
```

Yeah, something like this...would in fact be a requirement in a learning system or close to it. Neural nets do not come even close.

6:25pm. https://www.metalevel.at/prolog/dcg

Let me stop here. I've spent more time thinking about Spiral's top-down type system than the video here.

I am making very slow progress, but I'll do this for as long as it keeps my interest. After that I'll get back to studying game theory. I just want to get this curiosity sated."