"1:55pm. Just a bit more and I will resume.
https://re-library.com/translations/demon-sword-maiden/volume-1-rising-moon/chapter-11-flowery-night/

I almost ashamed to admit that I am reading this trash. I do admire the Chinese grit and murderhobo spirit, but I do wish the stories had creativity to go with their prolificness.

2:05pm. Ok, let me finally start. I want to go over the two Prolog books. Then I will do some easy exercises as warmup. Then I will start going through the game theory course.

I'll do a little Prolog every day from here on.

2:10pm. Ok, based on the first chapter, the book by Bratko is good. It starts out with examples right from the thick of it.

2:15pm. "An introduction to the first-order predicate calculus and resolution-based theorem proving can be found in several general books on artificial intelligence (Genesereth and Nilsson 1987; Ginsberg 1993; Poole et al. 1998; Russell and Norvig 2010; see also Flach 1994). Mathematical questions regarding the properties of Prolog's procedural meaning with respect to logic are analysed by Lloyd (1991)."

I should have the book by Russel.

4pm. Let me go do the chores. Enough of this for now.

I've skimmed half of Bratko's book. It will make for a good reference in the future. I am definitely intrigued by Prolog so far.

4:30pm. Let me get back to the book.

I guess I won't be doing the exercises today. Let me finish the books today and I will do the more fun stuff tomorrow. Sheesh.

5:10pm. Ok, let me stop here. Today I've been 80% thinking of how to infer higher kinded types, typeclasses and typecase and 20% reading the Prolog book.

I am constantly going back and forth between design decisions.

5:15pm. I am settled on wanting to have inl and join points, but I think I am going to skip on having values in the type system. I might put in higher ranked types though.

5:25pm. Let me call it a day here. I do not really feel like doing Prolog exercises now.

Actually, I still want to do just a bit more. I am more interested how Prolog works under the hood than anything else now. I am not too interested in programming in it, though I'd like to get familiar with the style and will expend some effort to get up to speed in that.

Let me take a look at the WAM book.

5:55pm. "Our style of teaching the WAM makes a special effort to consider carefully each feature of the WAM design in isolation by introducing separately and incrementally distinct aspects of Prolog. This allows us to explain as limpidly as possible specific principles proper to each. We then stitch and merge the different patches into larger pieces, introducing independent optimizations one at a time, converging eventually to the complete WAM design as described in [War83] or as overviewed in [War88]. Thus, in Chapter 2, we consider unification alone. Then, we look at flat resolution (that is, Prolog without backtracking) in Chapter 3. Following that, we turn to disjunctive definitions and backtracking in Chapter 4. At that point, we will have a complete, albeit naive, design for pure Prolog."

It sounds like the first 4 chapters of this is what I would be interested in. The pure subset of Prolog is the one I am most interested in.

6pm. Let me stop here. I can't tell whether I am being lazy or hard working. Somehow I am really enthused about designing the type system, but not so much about Prolog in particular.

From what I tell, the type system will be pretty vanilla. I can't really go a step beyond what currently exists.

At most, the big innovation will be introducing the typecase in an efficient manner, and combining it with Spiral's partial evaluation.

I'll weaken Spiral and get a somewhat plain language (in the vein of ML) as a result, but it should be significantly easier to program in than Spiral itself. At the very least, I won't have to spend hours fixing type errors after making a program. And I will have nearly all of the practical benefits even if not all of the power.

6:05pm. I do not really need layout type and values at the type level and all that complicated stuff. Forget about statically checked dimension sizes. I can just leave that to the partial evaluator and inline everything if it becomes necessary.

What I absolutely need are those tensors. And the GPU backend.

6:10pm. Doing that parser combinator library was a real pain in v0.09, but in v0.2 it will be just as easy as in F#. And as everything will be CPS'd and inlined in the right places, it should be extremely performant.

6:50pm. And that would be great if I actually needed performance or the GPU backend.

Still I should be able to implement unification with backtracking for type inference.

But it hardly a priority.

Let me end this thread here.

Yeah, languages do not matter.

I am going to make a decision. If I have to use Spiral again, I am going to bite the bullet and just put in top down typing. I'll do it rough and dirty and it will work great.

I have a great idea how to do the typecase. It will not be as elegant as typeclasses, but it sure as hell will be effective. It will give me everything that I need.

...Actually, no. Using typecase for everything would result in ridiculous type terms. Typeclass constraints have the great benefit of being able to be written down. So I am going to go with both. But neither of them will be difficult.

6:55pm. Languages really are worth it. I decided to go with Python back in 2015 which any sane person would do, I would not have developed all the techniques that I have now.

Spiral's partial evaluation really was worth taking the time to make. If I want to put an extensional type system on top of it, then I could not have asked for a finer foundation.

7pm. Somebody starting from the position of doing types first would not have arrived at it.

7:05pm. Ah, damn it. I wish I had a year of time to get the language design right. But I need to do what I must. Spiral was an absolute priority, but extensional type system for it is the cake. If I decide to do this, the universe won't wait for me.

7:10pm. I'll play with Prolog for a while just to get a concrete feel, but otherwise, I am done with this. It was a fun little excursion away from game theory, but I need to get back into the battle.

F# is perfectly enough for my purposes and so would be any SML-family language.

Now figure out ML and then we'll be cooking."