"9:50am. I got up 30m ago. Let me chill for a while and then I will start.
10am. https://j-novel.club/c/her-majesty-s-swarm-volume-1-part-1/

I saw this in yesterday's isekai thread and it seemed like an interesting premise. Let me read this and then I will start.

I really slept long today.

10:25am. Yeah, this is interesting enough to keep my attention. It could be good.

Yesterday I went through 30 chapters of Levelmaker and it is bad enough that I feel like apologizing to Demon Sword Maiden for calling it trash.

10:30am. Well, enough of that. It is time to start.

I had a lot of time to think about how I wanted to do convex sets. Without a doubt, it is going to be a pain to do the next part, but I need to do my best.

Yesterday the plan was to study game theory, but that is not what I need. I've gotten all that I can from lectures over the past 4.5 years. What I need to do is just start formalizing.

That is the one thing that will upgrade my skills.

Yesterday when I started on the paper, I did not expect to get anywhere, but I did figure out the definition. I did do something I thought I would not be able to. The conclusion is that I just need to focus and be very careful as I go through the paper.

10:35am. Eventually if I keep doing this, I will gain enough experience for formal papers to become easy.

Randomized testing is a good choice of focus. I thought of it being nothing but a gimmick, but the reality is that math is orders of magnitude less complex than programming.

10:40am. It is an enormous embarrassment to me that as an expert programmer, I cannot easily do something that should be 100 times easier than the main thing I am good at.

I should be able to become much, much better at this.

Formal proving is an enormous red herring. 'Dependent types are the future' my ass. They just waste my time and from a language design standpoint give me significant design issues that I cannot deal with.

Though I did gain something from doing theorem proving for so long, I am far below the level that I wanted to be. If you do something for months you'd expect to get good at it, but my productivity is rock bottom when it comes to Agda. I never went far with Lean either.

I did that insert function, but the effort I put in for that trivial thing is far out of proportion with the difficulty of the task. This is an indication that I won't be able to use it for things that actually matter.

10:50am. Ok. To start things out, let me do the convexity check for arrays. The condition will be that the second derivative be always positive or always negative (or neutral).

11:15am.

```
let error_bound_for_floats = 10.0 ** -7.0
let (=~) a b = abs (a - b) <= error_bound_for_floats
let (<=~) a b = a <= b + error_bound_for_floats
let (<~) a b = a < b + error_bound_for_floats
let (>=~) a b = a >= b - error_bound_for_floats
let (>~) a b = a > b - error_bound_for_floats

let prob_dist x = Array.forall ((<=) 0.0) x && Array.sum x =~ 1.0

let is_convex_set (l : float []) =
    if Array.length l < 2 then true
    else
        let rec loop f s i = if i < l.Length then let b, s = f s l.[i] in b && loop f s (i+1) else true
        let x = l.[1]
        let d = x - l.[0]
        let start op =
            loop (fun (d,x) x' ->
                let d' = x' - x
                let a = d' - d
                op a, (d',x')
                ) (d,x) 2

        start ((<=~) 0.0) || start ((>=~) 0.0)

is_convex_set [|0.0;0.5;0.7;0.9;1.0;1.1;1.15;1.17;1.18;1.0|]
```

This should suffice. Now what is the next step?

Let me take a look at the paper.

`probability distribution over a closed bounded convex set`

I need to figure out what this means.

12:20pm. The ideas are coming to me.

I am really quite engrossed in this.

I think I understand the issue at hand. But it is in fact quite difficult to define a probability distribution over a convex set.

12:25pm. The way I've defined a convex set is an array of points, but in actuality it is a geometric object. I forgot what Wildberger called it - some kind of connected spline.

So a probability distribution would have to be over that area.

It could be a bunch of points (with mass) over that area. Or it could be triangles (with mass) over that area.

I've been thinking that the canonical representation of probability distributions should in fact be triangular for a while now.

Still, something like that would be quite difficult to implement.

i'd really need to go through Wildberger's course, implementing the stuff as it comes.

By far the easiest choice is to have the probability distribution be made of points. And have those points be over the array elements.

12:30pm. It is just difficult to do it any other way.

12:35pm. Let me stop here. I know I started late, but I need to have that breakfast."