"https://wwwf.imperial.ac.uk/~dturaev/Hart0.pdf A Simple Adaptive Procedure Leading To Correlated Equilibrium
Here is the original regret matching paper. Maybe I will have some luck with this.

3:35pm. I had no idea that regret matching converges to a correlated equilibrium. Really?

5:05pm. Damn it, this is so annoying.

I've gone through that paper by Hart and now I am thinking. I a bit depressed that I cannot do the paper by Blackwell. To really feel safe, I should be able to do anything, but the mathematics in these papers is beyond me.

I might be able to do parts of the paper by Hart.

5:30pm. Done with lunch.

6:40pm. I am just going back and forth over the paper by Hart.

FFFFFFFFFFFFFFFFFFFFFFUCK!

I just do not fucking know!

For a minute there I thought I had made a breakthrough with randomized testing, but I can't follow the theorems in the papers by Hart and Blackwell at all.

For all I know, the stuff in these papers might as well be complete gibberish. I am actually depressed right now.

Just what can I do here? Do I dare pester Hart? Blackwell is long gone by this point, so obviously asking him is not an option.

7:05pm. Different people have different levels of understanding. My impression of Lanctot is that he himself did not have an exact grasp on the proofs I was asking him about for example. He is more of a programming guy like me.

...Yeah, I am going to have to ask. And if I do not get an answer, I am out of luck. But if I can get Hart to help my formalization effort, I can be sure that I will be getting insight from somebody who really knows his stuff. In return, the most I can give him is an actual formalization of the paper.

7:10pm. Getting research in the know to share the insight in return for formalizing their results seems like it would be a good trade. I really could learn that way. And it might interest the other side for whatever reason.

But let me leave that for later. I have to go through the other two references as well. No doubt they will crush me too, but only at that point will I be ready to give asking around a try.

7:20pm. If the doors close on me, I am going to go back to being an algorithms jockey and go all-in - even more so than I already am, on computer science. Maybe I'll make Spiral v0.2 as a gift to programming.

But in that case it would be depressing. It would be like an admission that I cannot cause the Singularity.

There really is a chance that I could do it if I can gather all the pieces of understanding together. How shameful is it to work this hard and as further insult have to rely on other people to get there. I will never live it down. It will be in this journal for others to laugh at for all eternity.

7:30pm. Let me stop here for the day.

Proving theorems = zero productivity. This seems like a reasonable approximation in the real world.

It really shows how much I want this that I am willing to spend and waste so much of my valuable time on this.

7:35pm. I want it, damn it, I want it. I want to make AI. And not just make, but understand it as well."