"1:45pm. Let me resume. Somehow chores and breakfast always take me a while.
I need to focus on this. I need to grasp what Metamath is doing.

2:10pm. "Brilliant mathematicians have discovered almost unimaginably profound results that rank among the crowning intellectual achievements of mankind. However, there is a sense in which modern abstract mathematics is behind the times, stuck in an era before computers existed. While no one disputes the remarkable results that have been achieved, communicating these results in a precise way to the uninitiated is virtually impossible. To describe these results, a terse informal language is used which despite its elegance is very difficult to learn. This informal language is not imprecise, far from it, but rather it often has omitted detail and symbols with hidden context that are implicitly understood by an expert but few others. Extremely complex technical meanings are associated with innocent-sounding English words such as “compact” and “measurable” that barely hint at what is actually being said. Anyone who does not keep the precise technical meaning constantly in mind is bound to fail, and acquiring the ability to do this can be achieved only through much practice and hard work. Only the few who complete the painful learning experience can join the small in-group of pure mathematicians. The informal language effectively cuts off the true nature of their knowledge from most everyone else."

I can't imagine a better way to do an intro than this particular paragraph. Simply brilliant prose.

2:20pm. "Suppose you have no formal training in abstract mathematics. But popular books you’ve read offer tempting glimpses of this world filled with profound ideas that have stirred the human spirit. You are not satisfied with the informal, watered-down descriptions you’ve read but feel it is important to grasp the underlying mathematics itself to understand its true meaning. It’s not practical to go back to school to learn it, though; you don’t want to dedicate years of your life to it. There are many important things in life, and you have to set priorities for what’s important to you. What would happen if you tried to pursue it on your own, in your spare time?

After all, you were able to learn a computer programming language such as Pascal on your own without too much difficulty, even though you had no formal training in computers. You don’t claim to be an expert in software design, but you can write a passable program when necessary to suit your needs. Even more important, you know that you can look at anyone else’s Pascal program, no matter how complex, and with enough patience figure out exactly how it works, even though you are not a specialist. Pascal allows you do anything that a computer can do, at least in principle. Thus you know you have the ability, in principle, to follow anything that a computer program can do: you just have to break it down into small enough pieces.

Here’s an imaginary scenario of what might happen if you naively adopted this same view of abstract mathematics and tried to pick it up on your own, in a period of time comparable to, say, learning a computer programming language."

The book really speaks to my heart.

2:45pm. 32/247.

"Even quite basic theorems would demand almost unbelievably vast books to display their proofs."
"Oh, of course no one ever really does it. It would take forever! You just show that you could do it, that’s sufficient."
"There is a theorem in the primitive notation of set theory that corresponds to the arithmetic theorem ‘1000 + 2000 = 3000’. The formula would be forbiddingly long. . . even if [one] knows the definitions and is asked to simplify the long formula according to them, chances are he will make errors and arrive at some incorrect result."
"The Principia Mathematica was the crowning achievement of the formalists. It was also the deathblow of the formalist view.. . . [Russell] failed, in three enormous volumes, to get beyond the elementary facts of arithmetic. He showed what can be done in principle and what cannot be done in practice. If the mathematical process were really one of strict, logical progression, we would still be counting our fingers.. . . One theoretician estimates. . . that a demonstration of one of Ramanujan’s conjectures assuming set theory and elementary analysis would take about two thousand pages; the length of a deduction from first principles is nearly inconceivable. . . The probabilists argue that. . . any very long proof can at best be viewed as only probably correct. . ."

3:20pm. I am really particularly tired right now. I really did overdo it last night.

"A good overview of higher-level proof verification languages (such as lcf and hol) is given in [22]. All of these languages are fundamentally different from Metamath in that much of the mathematical foundational knowledge is embedded in the underlying proof-verification program, rather than placed directly in the database that is being verified. These can have a steep learning curve for those without a mathematical background. For example, one usually must have a fair understanding of mathematical logic in order to follow their proofs."

Well, I'll last for 2 hours more.

4:50pm. 70/247. I really am too tired for this.

I am just skimming at this point. I think I'll do that for a while more and only tomorrow will I give the system a serious try.

5pm. 77/247.

Finally at chapter 3. I did not get the last chapter. I am going to have to experiment to get a feel for it.

6:05pm. 95/247. I'd be lying if I said I was paying attention at this point.

Let me just skim the rest to see what it is about and I will call it a day.

Tomorrow, I will see whether I can get some of the examples running.

6:15pm. 129/247.
"Thus mathematics may be defined as the subject in which we never know what we are talking about, nor whether what we are saying is true." - Bertrand Russell

This made me lol.

6:35pm. 157/247. Metamath feels like it would be an absolute pain in the ass to program in.

Since the book mentioned the thing would be like only 300 lines of Python, it might be interesting to look at the source.

Ok, that is enough of that for the day.

Somehow it is only at this point that I've stopped feeling drowsy.

6:40pm. I am not really sure whether I am on the right track with Metamath. But I do not need to learn some set theory.

At this point I am really thinking that all the mathematics departments should be razed to the ground. I mean it took me 5 days for that shitty theorem 3 from the CFR paper (not count the time spent messing with Lean and Agda.) From the outside math has a meritocratic feel, but now that I look it more closely it is exactly the opposite. It is nepotist at its core.

6:45pm. Well, it is fine. I've only been doing it for like 6 months. I'll get through set theory and build up my foundation.

I can only bemoan my fate for the world being in this state. Nonetheless, this is an obstacle I need to overcome if I want to meet my goal. I've given up on life for the sake of devouring the universe.

It is too much to ask that it just serves itself to me on a platter.

I am insufficiently determined. A part of me still has the mentality that I need to just rush through this in order to get back to programming.

The right way to go here is to take it slowly, even if it takes me a few years to wrap it up. Only then will I be able to do serious AI work.

It is fine if I struggle with these papers for a while. It won't last for too long - from a long term perspective.

My general programming skill is quite high. I just need to grind my maths to the same level and miracles will start happening one after another."