"2:50pm. The documentation in `set.mm` is good.
3:05pm. Or not.

```
  ${
    notnotrd.1 $e |- ( ph -> -. -. ps ) $.
    $( Deduction converting double-negation into the original wff, aka the
       double negation rule.  A translation of natural deduction rule ` -. -. `
       -C, ` _G |- -. -. ps ` => ` _G |- ps ` ; see ~ natded .  This is
       definition NNC in [Pfenning] p. 17.  This rule is valid in classical
       logic (which MPE uses), but not intuitionistic logic.  (Contributed by
       DAW, 8-Feb-2017.) $)
    notnotrd $p |- ( ph -> ps ) $=
      ( wn notnot2 syl ) ABDDBCBEF $.
  $}
```

All the examples are like this.

I am not too much of a fan of this. I am going to have to look into Metamath's source.

It is in C. No forget this. Seriously.

Metamath might be Spartan, but that just means it has to use syntax tricks everywhere rather than having a properly designed language. Despite reading this far, I am still not sure how in chapter two where it uses the reverse polish notation it goes to actually referencing past steps.

All this is not very appealing to me. I want something easier to get into. My goal right now is just to study set theory. If I have to study how the language itself is made from the ground up then that is much less useful to me.

3:55pm. Had to take a break. I needed to think anyway. I am not really what I should do here.

Let me check out Mizar and HOL Light and I will think of what I want to do after that.

I am thinking that as my intro to Metamath, I should reimplement the prover in F#.

It is only around 350 lines of Python. I could get that, and then translate it to F#.

https://www.cl.cam.ac.uk/~jrh13/hol-light/
http://mizar.org/project/bibliography.html

Let me start with Mizar.

4:20pm. Ok, I've looked into Mizar for a bit.

It looks like a regular language. That is all I can comment on it.

I am not really interest in diving further into it.

Let me take a look at HOL Light next.

5:10pm. So far I've been reading the appendix of the HOL Light manual, on the history and evolution of HOL Light. It might be worth looking into at some point.

5:25pm. http://us.metamath.org/downloads/mmverify.py

Here is that 350 lines Python program.

http://us.metamath.org/other.html#verifiers

There are a bunch of them here including Scala, Haskell and C# versions.

I might want to look at the last two.

5:35pm. But not today. If I were to say how I feel right now, it would be like some sort of fly.

I keep going from thing to thing without ever making any real progress in what I really want.

5:40pm. While investigating HOL Light might be interesting at some point as it seems to have developed libraries in various areas, it is beyond the point at this moment. The same goes for Mizar.

What I need to do is study set theory and diving into a library like Metamath makes the most sense. And I agree with Metamath philosophy of using substitutions for this sort of thing.

With the fields of mathematics being in the state that they are, computability of Coq and Agda is actually a burden to me.

It is ironic that I need to shed those beliefs and do what I hate before I can really truly reject them. Know your enemy and all that.

5:45pm. Here is what my goal for tomorrow will be.

I am going to take that example from chapter two, get it to run with the standard interpreter. Then I am going to take the C# one and begin translating it to F#. This should not me take longer than a few days at most.

5:50pm. Really, I can accept mathematics being just symbol pushing just fine.

I need to get better, better and better at this. At all the logics. I need to take in everything that math has to offer. Then I will devour the game theory papers that are giving me so much trouble whole.

When I am able to follow any kind of math that pertains to my interest, then my rise will begin.

5:55pm. I feel that the randomized test I did for the CFR proof was something special, but I won't be able to make full use of the technique until I get my fundamentals in order.

6pm. I am done for the day here."